Move 'CopyDirectoryContents' into the new directory utilities folder
