+++ src/DoomerPublish/DoomerPublishTool/Service/PublisherService/PublishTasks/CopyProjectToTempDirTask.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using DoomerPublish.Utils;
+using Microsoft.Extensions.Logging;
+using System.Diagnostics;
 
 namespace DoomerPublish.PublishTasks;
 
@@ -44,51 +46,14 @@ internal sealed class CopyProjectToTempDirTask : IPublishTask
 		// Create base folder.
 		_ = Directory.CreateDirectory(output);
 
-		this._logger.LogDebug("Start copy to temporary directory.", projectContext.ProjectName);
+		this._logger.LogDebug("Start copy '{FolderPath}' to temporary directory.", projectContext.ProjectPath);
 
-		this.CopyDirectoryContents(projectContext.ProjectPath, output);
+		DirectoryUtils.CopyDirectoryContents(projectContext.ProjectPath, output);
 
 		// The project path must be changed.
 		projectContext.ProjectPath = output;
 		this._logger.LogInformation("Copied project over to {TempDirLocation}.", output);
 
 		return Task.CompletedTask;
-	}
-
-	// Recursively copies all the directories and their child directories into the output folder.
-	// This will ensure all files persist, even if they are being locked by another operation.
-	private void CopyDirectoryContents(string sourceFolder, string outputFolder)
-	{
-		if (!Directory.Exists(outputFolder))
-		{
-			_ = Directory.CreateDirectory(outputFolder);
-		}
-
-		this._logger.LogDebug("Copying files from folder {FolderPath}.", sourceFolder);
-
-		// TODO: Look into a faster way of doing this. Especially bigger mods like Floppy Disk Mod just takes so long doing this.
-		foreach (var filePath in Directory.GetFiles(sourceFolder))
-		{
-			var fileName = Path.GetFileName(filePath);
-			var outputPath = Path.Combine(outputFolder, fileName);
-
-			using var inputFile = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
-			using var outputFile = new FileStream(outputPath, FileMode.Create);
-			var buffer = new byte[0x10000];
-			int bytes;
-
-			while ((bytes = inputFile.Read(buffer, 0, buffer.Length)) > 0)
-			{
-				outputFile.Write(buffer, 0, bytes);
-			}
-		}
-
-		foreach (var directoryPath in Directory.GetDirectories(sourceFolder))
-		{
-			var outputSubFolder = Path.Combine(outputFolder, Path.GetFileName(directoryPath));
-			_ = Directory.CreateDirectory(outputSubFolder);
-
-			this.CopyDirectoryContents(directoryPath, outputSubFolder);
-		}
 	}
 }
